move abstract rector traits to AbstractRector namespace
