+++ src/Rector/AbstractRector/ValueResolverTrait.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Rector;
+namespace Rector\Rector\AbstractRector;
 
 use PhpParser\Node\Expr;
 use Rector\PhpParser\Node\Value\ValueResolver;
